Update php driver
Thanks for this AWESOME list. I've updated the PHP driver to point to GraphAware as it was acquired since then. Also, currently I point to the v4.0 branch (still in beta because of the Bolt support which isn't officially stable by Neo4j), but even for http it is now the recommended version to use.

Cheers m8