NodeRank is superseded by Neo4j Graph Algorithms (#16)
According to NodeRank  github README, "As of October 17th 2017, this module is retiring. This means it will no longer be maintained and released together with new versions of the GraphAware Framework and Neo4j. The last compatible Neo4j version is 3.2.5. NodeRank is superseded by a much more active Neo4j Graph Algorithms project."